Sound can be disabled for a specific time

+++ include/StopWatch.h
@@ -13,7 +13,7 @@ public:
      * @param stop The later time point
      * @return difference in seconds
      */
-    float time_difference( const timespec& start, const timespec& stop ) {
+    double time_difference( const timespec& start, const timespec& stop ) {
         float time = stop.tv_sec - start.tv_sec + (stop.tv_nsec - start.tv_nsec) * 1E-9;
         return time;
     }
@@ -28,7 +28,7 @@ public:
 
     void Start() { clock_gettime(CLOCK_MONOTONIC, &_start); }
     void Stop() { clock_gettime(CLOCK_MONOTONIC, &_stop);}
-    float TimeElapsed();
+    double TimeElapsed();
 
 };
